bug(/review): Fix agent not posting feedback in GitHub (#141)
This change modifies the prompt to constrain the agent on how the review
must be submitted to GitHub. This aims to improve what is seen in #134
where "more than 50% of the time PR review fails due to tool problems
posting the review."

Since this is occurring intermittently - this is a probabilistic issue,
not a permissions issue, which is why the prompt is being changed.

Primary changes: 
1. Sometimes the agent made tool calls for a fake repo `owner` to make
tool calls. The change instructs the agent to list the repository owner.
1. Surprisingly, even when the agent hallucinated a fake repo `owner`
and made tool calls with this, our logs show that these calls were
successful (so the agent thought things were working), which is why
workflow errors did not surface. I have filed a
[bug](https://github.com/github/github-mcp-server/issues/842) with
GitHub MCP to ensure they are returning failures, and will ensure that
`gemini-cli` logs/telemetry is properly capturing and recording tool
call failures.
1. Sometimes the agent didn't attempt to make tool calls. The prompt was
made to be more explicit that this is a necessity.

Fixes #134