feat: support /fix slash command and workflow (#190)
### TL;DR
Add workflow for `@gemini-cli /fix` to fix a GitHub Issue

### Design
This workflow seeks to take similar approaches to other workflows, such
as `gemini-cli.yml` or `gemini-pr-review.yml`, but differs primarily in
its prompt. The differences aim to enhance the reliability/performance
of the workflow by consistently adopting XML tags, removing newline
characters that are used solely for breaking up long lines in IDEs, and
placing a greater emphasis on providing examples that use least-to-most
prompting.

### Testing
GitHub actions are difficult to test before they are in `main`, as the
workflow does not exist at
https://github.com/google-github-actions/run-gemini-cli/actions to
invoke. Therefore, this change only adds the workflow internally for
evaluation/iteration to enable testing, and when it is provably useful
it can be added to the externally facing workflows in
`/examples/workflows/`.